requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
requests
beautifulsoup4